web/requirements.txt
Django==1.11.18
Markdown==3.0.1
coverage==4.5.2
django-braces==1.13.0
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
djangorestframework==3.9.1
django-shotgun==0.2
drf-extensions==0.4.0
drf-hal-json==0.9.1
drf-nested-fields==0.9.4
factory-boy==2.11.1
Faker==1.0.1
psycopg2-binary==2.7.6.1
pyslack-real==0.6.0
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
sphinx-me==0.3
sqlparse==0.2.4
tqdm==4.29.1
urllib3==1.24.1
